"Final Version" Bugs/shortcomings include:
The yearDonated column remains blank on the searchResultsPage

The yearDonated value is blank on the Individual result page
("TreePage")

Quadrant image fails to load in the main page

The first search may not work as the data will not have initialized; a
second search works as the data has updated 

Some maps may or may not work; their markers and centering are off.

The Search results page does not initialize with the data from the
previous search parameter 

The Search Results page does not contain a "Search for
non-donated objects" capability

The application lacks a method of regular updating; this is a problem as
Cindy wanted to run the application 24/7 and it only updates on load
right now.